Refactor manfifest integration tests into runtime and planning tests.
Also splits out the shared setup function.
